Counts of attribute balance

This function prints out a summary of the counts of each level for each attribute across all choice questions as well as the two-way counts across all pairs of attributes for a given design.

cbc_balance(design, atts = NULL)

Arguments

design

A data frame of a survey design.

atts

A vector of attributes. Defaults to NULL, in which case all of the column names of the design data frame are used except for those that end in "ID".

Value

A list showing counts of the number of choice questions that contain the unique number of levels for each attribute.

Examples

library(cbcTools)

# A simple conjoint experiment about apples

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c('Poor', 'Average', 'Excellent')
)

# Make a randomized survey design
design <- cbc_design(
  profiles = profiles,
  n_resp   = 300, # Number of respondents
  n_alts   = 3, # Number of alternatives per question
  n_q      = 6 # Number of questions per respondent
)

# Inspect the design balance
cbc_balance(design)
#> ==============================
#> Attribute counts:
#> 
#> price:
#> 
#>   1 1.5   2 2.5   3 3.5   4 4.5   5 
#> 618 563 557 589 599 638 653 606 577 
#> 
#> type:
#> 
#>       Fuji       Gala Honeycrisp 
#>       1792       1806       1802 
#> 
#> freshness:
#> 
#>      Poor   Average Excellent 
#>      1846      1787      1767 
#> 
#> ==============================
#> Pairwise attribute counts:
#> 
#> price & type:
#>      
#>       Fuji Gala Honeycrisp
#>   1    209  219        190
#>   1.5  195  179        189
#>   2    179  199        179
#>   2.5  204  184        201
#>   3    197  202        200
#>   3.5  223  203        212
#>   4    199  230        224
#>   4.5  187  191        228
#>   5    199  199        179
#> 
#> price & freshness:
#>      
#>       Poor Average Excellent
#>   1    206     196       216
#>   1.5  212     179       172
#>   2    179     187       191
#>   2.5  204     219       166
#>   3    187     195       217
#>   3.5  222     223       193
#>   4    227     203       223
#>   4.5  208     196       202
#>   5    201     189       187
#> 
#> type & freshness:
#>             
#>              Poor Average Excellent
#>   Fuji        626     575       591
#>   Gala        618     586       602
#>   Honeycrisp  602     626       574
#> 

# Inspect the design overlap
cbc_overlap(design)
#> ==============================
#> Counts of attribute overlap:
#> (# of questions with N unique levels)
#> 
#> price:
#> 
#>    1    2    3 
#>   14  493 1293 
#> 
#> type:
#> 
#>    1    2    3 
#>  174 1222  404 
#> 
#> freshness:
#> 
#>    1    2    3 
#>  176 1181  443 
#>