Counts of attribute balance

This function prints out a summary of the counts of each level for each attribute across all choice questions as well as the two-way counts across all pairs of attributes for a given design.

cbc_balance(design, atts = NULL)

Arguments

design

A data frame of a survey design.

atts

A vector of attributes. Defaults to NULL, in which case all of the column names of the design data frame are used except for those that end in "ID".

Value

A list showing counts of the number of choice questions that contain the unique number of levels for each attribute.

Examples

library(cbcTools)

# A simple conjoint experiment about apples

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c('Poor', 'Average', 'Excellent')
)

# Make a randomized survey design
design <- cbc_design(
  profiles = profiles,
  n_resp   = 300, # Number of respondents
  n_alts   = 3, # Number of alternatives per question
  n_q      = 6 # Number of questions per respondent
)

# Inspect the design balance
cbc_balance(design)
#> ==============================
#> price x type 
#> 
#>         Fuji Gala Honeycrisp
#>      NA 1767 1818       1815
#> 1   603  215  210        178
#> 1.5 602  189  208        205
#> 2   604  205  196        203
#> 2.5 581  187  191        203
#> 3   613  213  194        206
#> 3.5 578  164  204        210
#> 4   634  195  218        221
#> 4.5 563  185  201        177
#> 5   622  214  196        212
#> 
#> price x freshness 
#> 
#>         Poor Average Excellent
#>      NA 1788    1817      1795
#> 1   603  193     205       205
#> 1.5 602  222     203       177
#> 2   604  206     203       195
#> 2.5 581  181     186       214
#> 3   613  192     202       219
#> 3.5 578  193     212       173
#> 4   634  201     217       216
#> 4.5 563  186     201       176
#> 5   622  214     188       220
#> 
#> type x freshness 
#> 
#>                 Poor Average Excellent
#>              NA 1788    1817      1795
#> Fuji       1767  605     606       556
#> Gala       1818  560     605       653
#> Honeycrisp 1815  623     606       586
#> 

# Inspect the design overlap
cbc_overlap(design)
#> ==============================
#> Counts of attribute overlap:
#> (# of questions with N unique levels)
#> 
#> price:
#> 
#>    1    2    3 
#>   19  519 1262 
#> 
#> type:
#> 
#>    1    2    3 
#>  204 1165  431 
#> 
#> freshness:
#> 
#>    1    2    3 
#>  178 1182  440 
#>