Counts of attribute balance

This function prints out a summary of the individual and pairwise counts of each level for each attribute across all choice questions in the design.

cbc_balance(design)

Arguments

design

A data frame of a survey design.

Value

Prints the individual and pairwise counts of the number of times each attribute levels in shown in the design.

Examples

library(cbcTools)

# A simple conjoint experiment about apples

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c('Poor', 'Average', 'Excellent')
)

# Make a randomized survey design
design <- cbc_design(
  profiles = profiles,
  n_resp   = 300, # Number of respondents
  n_alts   = 3, # Number of alternatives per question
  n_q      = 6 # Number of questions per respondent
)

# Inspect the design balance
cbc_balance(design)
#> ==============================
#> price x type 
#> 
#>         Fuji Gala Honeycrisp
#>      NA 1768 1858       1774
#> 1   600  198  219        183
#> 1.5 610  206  205        199
#> 2   551  174  206        171
#> 2.5 611  182  203        226
#> 3   610  198  201        211
#> 3.5 627  210  195        222
#> 4   558  182  197        179
#> 4.5 619  206  217        196
#> 5   614  212  215        187
#> 
#> price x freshness 
#> 
#>         Poor Average Excellent
#>      NA 1847    1851      1702
#> 1   600  193     220       187
#> 1.5 610  186     221       203
#> 2   551  195     170       186
#> 2.5 611  208     216       187
#> 3   610  219     198       193
#> 3.5 627  222     214       191
#> 4   558  188     195       175
#> 4.5 619  236     206       177
#> 5   614  200     211       203
#> 
#> type x freshness 
#> 
#>                 Poor Average Excellent
#>              NA 1847    1851      1702
#> Fuji       1768  620     595       553
#> Gala       1858  637     638       583
#> Honeycrisp 1774  590     618       566
#> 

# Inspect the design overlap
cbc_overlap(design)
#> ==============================
#> Counts of attribute overlap:
#> (# of questions with N unique levels)
#> 
#> price:
#> 
#>    1    2    3 
#>   16  510 1274 
#> 
#> type:
#> 
#>    1    2    3 
#>  163 1204  433 
#> 
#> freshness:
#> 
#>    1    2    3 
#>  197 1198  405 
#>