Counts of attribute balance

This function prints out a summary of the counts of each level for each attribute across all choice questions as well as the two-way counts across all pairs of attributes for a given design.

cbc_balance(design, atts = NULL)

Arguments

design

A data frame of a survey design.

atts

A vector of attributes. Defaults to NULL, in which case all of the column names of the design data frame are used except for those that end in "ID".

Value

A list showing counts of the number of choice questions that contain the unique number of levels for each attribute.

Examples

library(cbcTools)

# A simple conjoint experiment about apples

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  freshness = c("Excellent", "Average", "Poor"),
  type      = c("Fuji", "Gala", "Honeycrisp")
)

# Make a randomized survey design
design <- cbc_design(
  profiles = profiles,
  n_resp   = 300, # Number of respondents
  n_alts   = 3, # Number of alternatives per question
  n_q      = 6 # Number of questions per respondent
)

# Inspect the design balance
cbc_balance(design)
#> ==============================
#> Attribute counts:
#> 
#> price:
#> 
#>   1 1.5   2 2.5   3 3.5   4 4.5   5 
#> 603 616 571 594 607 577 626 574 632 
#> 
#> freshness:
#> 
#> Excellent   Average      Poor 
#>      1800      1825      1775 
#> 
#> type:
#> 
#>       Fuji       Gala Honeycrisp 
#>       1755       1808       1837 
#> 
#> ==============================
#> Pairwise attribute counts:
#> 
#> price & freshness:
#>      
#>       Excellent Average Poor
#>   1         213     181  209
#>   1.5       201     194  221
#>   2         200     211  160
#>   2.5       184     211  199
#>   3         203     197  207
#>   3.5       192     198  187
#>   4         203     237  186
#>   4.5       190     192  192
#>   5         214     204  214
#> 
#> price & type:
#>      
#>       Fuji Gala Honeycrisp
#>   1    191  216        196
#>   1.5  203  180        233
#>   2    199  187        185
#>   2.5  200  191        203
#>   3    205  207        195
#>   3.5  188  204        185
#>   4    195  209        222
#>   4.5  188  188        198
#>   5    186  226        220
#> 
#> freshness & type:
#>            
#>             Fuji Gala Honeycrisp
#>   Excellent  589  626        585
#>   Average    573  599        653
#>   Poor       593  583        599
#> 

# Inspect the design overlap
cbc_overlap(design)
#> ==============================
#> Counts of attribute overlap:
#> (# of questions with N unique levels)
#> 
#> price:
#> 
#>    1    2    3 
#>   17  496 1287 
#> 
#> freshness:
#> 
#>    1    2    3 
#>  182 1208  410 
#> 
#> type:
#> 
#>    1    2    3 
#>  199 1212  389 
#>