Make a random or Bayesian D-efficient choice-based conjoint survey design

This function creates a data frame containing a choice-based conjoint survey design where each row is an alternative. Designs can be either a randomized or Bayesian D-efficient, in which case an implementation of the CEA or Modfed Federov algorithm is used via the idefix package

cbc_design(
  profiles,
  n_resp,
  n_alts,
  n_q,
  n_blocks = 1,
  n_draws = 50,
  no_choice = FALSE,
  n_start = 5,
  label = NULL,
  priors = NULL,
  prior_no_choice = NULL,
  probs = FALSE,
  method = NULL,
  keep_db_error = FALSE,
  max_iter = 50,
  parallel = TRUE
)

Arguments

profiles

A data frame in which each row is a possible profile. This can be generated using the cbc_profiles() function.

n_resp

Number of survey respondents.

n_alts

Number of alternatives per choice question.

n_q

Number of questions per respondent.

n_blocks

Number of blocks used in Bayesian D-efficient design. Max allowable is one block per respondent, defaults to 1, meaning every respondent sees the same set of choice questions.

n_draws

Number of draws used in simulating the prior distribution used in Bayesian D-efficient designs. Defaults to 50.

no_choice

Include a "no choice" option in the choice sets? Defaults to FALSE. If TRUE, the total number of alternatives per question will be one more than the provided n_alts argument.

n_start

A numeric value indicating the number of random start designs to use in obtaining a Bayesian D-efficient design. The default is 5. Increasing n_start can result in a more efficient design at the expense of increased computational time.

label

The name of the variable to use in a "labeled" design (also called an "alternative-specific design") such that each set of alternatives contains one of each of the levels in the label attribute. Currently only compatible with randomized designs. If used, the n_alts argument will be ignored as its value is defined by the unique number of levels in the label variable. Defaults to NULL.

priors

A list of one or more assumed prior parameters used to generate a Bayesian D-efficient design. If NULL (the default), a randomized design will be generated.

prior_no_choice

Prior utility value for the "no choice" alternative. Only required if no_choice = TRUE. Defaults to NULL.

probs

If TRUE, for Bayesian D-efficient designs the resulting design includes average predicted probabilities for each alternative in each choice set given the sample from the prior preference distribution. Defaults to FALSE.

method

Which method to use for obtaining a Bayesian D-efficient design, "CEA" or "Modfed"? If priors are specified, it defaults to "CEA", otherwise it defaults to NULL. See ?idefix::CEA and ?idefix::Modfed for more details.

keep_db_error

If TRUE, for Bayesian D-efficient designs the returned object will be a list containing the design and the DB-error score. Defaults to FALSE.

max_iter

A numeric value indicating the maximum number allowed iterations when searching for a Bayesian D-efficient design. The default is 50.

parallel

Logical value indicating whether computations should be done over multiple cores. The default is TRUE.

Value

A data frame containing a choice-based conjoint survey design where each row is an alternative.

Examples

library(cbcTools)

# A simple conjoint experiment about apples

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c('Poor', 'Average', 'Excellent')
)

# Make a randomized survey design
design_rand <- cbc_design(
  profiles = profiles,
  n_resp   = 300, # Number of respondents
  n_alts   = 3,   # Number of alternatives per question
  n_q      = 6    # Number of questions per respondent
)

# Make a randomized survey design with a "no choice" option
design_rand_nochoice <- cbc_design(
  profiles  = profiles,
  n_resp    = 300, # Number of respondents
  n_alts    = 3,   # Number of alternatives per question
  n_q       = 6,   # Number of questions per respondent
  no_choice = TRUE
)

# Make a randomized labeled survey design with each "type" appearing in
# each choice question
design_rand_labeled <- cbc_design(
  profiles  = profiles,
  n_resp    = 300, # Number of respondents
  n_alts    = 3,   # Number of alternatives per question
  n_q       = 6,   # Number of questions per respondent
  label     = "type"
)

# Make a Bayesian D-efficient design with a prior model specified
# Note that by default parallel = TRUE.
design_deff <- cbc_design(
    profiles  = profiles,
    n_resp    = 300, # Number of respondents
    n_alts    = 3,  # Number of alternatives per question
    n_q       = 6,  # Number of questions per respondent
    n_start   = 1,
    priors = list(
        price     = -0.1,
        type      = c(0.1, 0.2),
        freshness = c(0.1, 0.2)
    ),
    method = "CEA",
    parallel = FALSE
)
#> Bayesian D-efficient design found with DB-error of 1.31223