Counts of attribute overlap

This function prints out a summary of the amount of "overlap" across attributes within the choice questions. For example, for each attribute, the count under "1" is the number of choice questions in which the same level was shown across all alternatives for that attribute (because there was only one level shown). Likewise, the count under "2" is the number of choice questions in which only two unique levels of that attribute were shown, and so on.

cbc_overlap(design)

Arguments

design

A data frame of a survey design.

Value

Prints the counts of the number of choice questions that contain the unique number of levels for each attribute.

Examples

library(cbcTools)

# A simple conjoint experiment about apples

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  freshness = c("Excellent", "Average", "Poor"),
  type      = c("Fuji", "Gala", "Honeycrisp")
)

# Make a randomized survey design
design <- cbc_design(
  profiles = profiles,
  n_resp   = 300, # Number of respondents
  n_alts   = 3, # Number of alternatives per question
  n_q      = 6 # Number of questions per respondent
)

# Inspect the design balance
cbc_balance(design)
#> =====================================
#> Individual attribute level counts
#> 
#> price:
#> 
#>   1 1.5   2 2.5   3 3.5   4 4.5   5 
#> 604 561 590 626 620 612 597 615 575 
#> 
#> freshness:
#> 
#> Excellent   Average      Poor 
#>      1801      1824      1775 
#> 
#> type:
#> 
#>       Fuji       Gala Honeycrisp 
#>       1807       1769       1824 
#> 
#> =====================================
#> Pairwise attribute level counts
#> 
#> price x freshness:
#> 
#>         Excellent Average Poor
#>      NA      1801    1824 1775
#> 1   604       190     226  188
#> 1.5 561       199     174  188
#> 2   590       200     207  183
#> 2.5 626       207     201  218
#> 3   620       198     213  209
#> 3.5 612       198     202  212
#> 4   597       194     199  204
#> 4.5 615       213     215  187
#> 5   575       202     187  186
#> 
#> price x type:
#> 
#>         Fuji Gala Honeycrisp
#>      NA 1807 1769       1824
#> 1   604  219  157        228
#> 1.5 561  165  200        196
#> 2   590  197  185        208
#> 2.5 626  204  218        204
#> 3   620  227  192        201
#> 3.5 612  202  222        188
#> 4   597  179  203        215
#> 4.5 615  213  196        206
#> 5   575  201  196        178
#> 
#> freshness x type:
#> 
#>                Fuji Gala Honeycrisp
#>             NA 1807 1769       1824
#> Excellent 1801  596  592        613
#> Average   1824  613  605        606
#> Poor      1775  598  572        605
#> 

# Inspect the design overlap
cbc_overlap(design)
#> ==============================
#> Counts of attribute overlap:
#> (# of questions with N unique levels)
#> 
#> price:
#> 
#>    1    2    3 
#>   20  486 1294 
#> 
#> freshness:
#> 
#>    1    2    3 
#>  180 1190  430 
#> 
#> type:
#> 
#>    1    2    3 
#>  186 1200  414 
#>